requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
minio==7.0.2
multidict==5.1.0
python-dotenv==0.15.0
requests==2.25.1
rfc3986==1.4.0
slack-sdk==3.3.2
sniffio==1.2.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
